User can submit self chat values
