+++ public/components/Chat.jsx
@@ -1,5 +1,8 @@
 import React from 'react';
+
 var RTM = require("satori-rtm-sdk");
+
+import ChatSubmit from './ChatSubmit.jsx';
 
 class Chat extends React.Component {
   constructor(props) {
@@ -9,36 +12,6 @@ class Chat extends React.Component {
     };
 
   }
-
-  submitMessage() {
-    var endpoint = "wss://open-data.api.satori.com";
-    var appkey = "B1c1F2BfAC806D6daFC3D4Fa4A4B00ED";
-    var role = "sacangelhack";
-    var roleSecretKey = "3EF46ECF5a1d1F365fdF37c0dA9e38d3";
-    var channel = "sacangelhack";
-    var self = this;
-
-    var roleSecretProvider = RTM.roleSecretAuthProvider(role, roleSecretKey);
-
-    var rtm = new RTM(endpoint, appkey, {
-      authProvider: roleSecretProvider,
-    });
-
-    var subscription = rtm.subscribe(channel, RTM.SubscriptionMode.SIMPLE);
-
-    /* publish a message after being subscribed to sync on subscription */
-    subscription.on('enter-subscribed', function () {
-      let message = {
-        test: "message 2"
-      }
-      rtm.publish(channel, message, function (pdu) {
-        console.log("Publish ack:", pdu);
-      });
-    });
-
-    rtm.start();
-  }
-
   render() {
     var endpoint = "wss://open-data.api.satori.com";
     var appkey = "B1c1F2BfAC806D6daFC3D4Fa4A4B00ED";
@@ -79,7 +52,7 @@ class Chat extends React.Component {
           })
         }
         <div>
-          <button onClick={this.submitMessage}>Test</button>
+          <ChatSubmit />
         </div>
       </div>
     );
